state: update inherited_vars and tracking dicts when adding vars
Cached vars were not updated when depending on dynamically added vars, because
the dependency tracking dicts update substate_var_dependencies based on whether
the var is inherited and inherited_vars was not being updated (and not being
updated recursively) when vars were added to the state.

Now when a var is added, push the var down into all substates as an inherited
var, and recursively (depth-first) update all substate
_computed_var_dependencies and _substate_var_dependencies in case any of them
were depending on a var that gets added later in the initialization process.